Webpacked (#140)
* Remove existing module.json

* Webpacked Version

Here we add some webpack bits. The package.json and lock file allow npm to install the required pieces.

The `webpack.config.js` provides the simple webpack config, this reduces the module down to a dingle 800kb ish js file. Hurrah!

The github workflow has been updated to create a module.json from the module-template.json (more on this later), install the npm components and move across the files (generated in `dist`) to the zip file.

The `module.json` has moved to `module-template.json` and this is the file used to run the module in bundled mode. The module-dev.json will run the module in unbundled mode. In your local dev environment copy the `module-dev.js` to `module.js`.

The `module.js` is now ignored by git.