pt-BR Translation (#276)
* pt-BR Translation

Add translation to Brazilian Portuguese. These translations are based on old oficial books of D&D5e, cause today there's no oficial translation and the majority of the community uses the community translation itself (or just play it in english). When a translation would not exists, I translate to the most faithful translation to pt-BR.

* Update pt-BR.json

* Update pt-BR.json

* Update pt-BR.json

* Update pt-BR.json

* Update pt-BR.json

* Update pt-BR.json

* Update pt-BR.json

* Update pt-BR.json